// src/app/shared/models/wine-region.model.ts
export enum WineRegion {
  BORDEAUX_LEFT_BANK = 'BORDEAUX_LEFT_BANK',
  BORDEAUX_RIGHT_BANK = 'BORDEAUX_RIGHT_BANK',
  BORDEAUX_SAUTERNES = 'BORDEAUX_SAUTERNES',
  BURGUNDY_COTE_DE_NUITS = 'BURGUNDY_COTE_DE_NUITS',
  BURGUNDY_COTE_DE_BEAUNE = 'BURGUNDY_COTE_DE_BEAUNE',
  BURGUNDY_WHITE = 'BURGUNDY_WHITE',
  BURGUNDY_BEAUJOLAIS = 'BURGUNDY_BEAUJOLAIS',
  CHAMPAGNE = 'CHAMPAGNE',
  ALSACE = 'ALSACE',
  RHONE_NORTH = 'RHONE_NORTH',
  RHONE_SOUTH = 'RHONE_SOUTH',
  LOIRE_WHITE = 'LOIRE_WHITE',
  LOIRE_RED = 'LOIRE_RED',
  LANGUEDOC_ROUSSILLON = 'LANGUEDOC_ROUSSILLON',
  JURA = 'JURA',
  RIOJA = 'RIOJA',
  RIBERA_DEL_DUERO = 'RIBERA_DEL_DUERO',
  BAROLO = 'BAROLO',
  BARBARESCO = 'BARBARESCO',
  TUSCANY = 'TUSCANY',
  OTHER = 'OTHER',
}

export const WINE_REGION_LABELS: Record<WineRegion, string> = {
  [WineRegion.BORDEAUX_LEFT_BANK]:    'Bordeaux Rive Gauche',
  [WineRegion.BORDEAUX_RIGHT_BANK]:   'Bordeaux Rive Droite',
  [WineRegion.BORDEAUX_SAUTERNES]:    'Bordeaux Sauternes',
  [WineRegion.BURGUNDY_COTE_DE_NUITS]:'Bourgogne Côte de Nuits',
  [WineRegion.BURGUNDY_COTE_DE_BEAUNE]:'Bourgogne Côte de Beaune',
  [WineRegion.BURGUNDY_WHITE]:        'Bourgogne Blanc',
  [WineRegion.BURGUNDY_BEAUJOLAIS]:   'Beaujolais',
  [WineRegion.CHAMPAGNE]:             'Champagne',
  [WineRegion.ALSACE]:                'Alsace',
  [WineRegion.RHONE_NORTH]:           'Rhône Nord',
  [WineRegion.RHONE_SOUTH]:           'Rhône Sud',
  [WineRegion.LOIRE_WHITE]:           'Loire Blanc',
  [WineRegion.LOIRE_RED]:             'Loire Rouge',
  [WineRegion.LANGUEDOC_ROUSSILLON]:  'Languedoc-Roussillon',
  [WineRegion.JURA]:                  'Jura',
  [WineRegion.RIOJA]:                 'Rioja',
  [WineRegion.RIBERA_DEL_DUERO]:      'Ribera del Duero',
  [WineRegion.BAROLO]:                'Barolo',
  [WineRegion.BARBARESCO]:            'Barbaresco',
  [WineRegion.TUSCANY]:               'Toscane',
  [WineRegion.OTHER]:                 'Autre',
};

export const WINE_REGION_OPTIONS = Object.values(WineRegion).map(value => ({
  value,
  label: WINE_REGION_LABELS[value],
}));